% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_helpers.R
\name{.get_parser_version}
\alias{.get_parser_version}
\title{Helper to retrieve the installed or development package version}
\usage{
.get_parser_version()
}
\description{
Helper to retrieve the installed or development package version
}
\keyword{internal}
