% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_helpers.R
\name{.split_plate_name}
\alias{.split_plate_name}
\title{Optional: split Plate Name into named components}
\usage{
.split_plate_name(metadata, name_components)
}
\arguments{
\item{metadata}{Single-row metadata data frame containing a \code{"Plate Name"}
column.}

\item{name_components}{Optional character vector of component names.}
}
\value{
Metadata data frame with additional component columns appended if
\code{name_components} is provided.
}
\description{
Optional: split Plate Name into named components
}
\keyword{internal}
